scenarios/yaml_load/requirements.txt
requests
pyyaml
pycgi
chalice
pandas